ADD: Access to the endpoint

+++ gigimovies/gigimovies/DataLayer/Entities/MovieEntity.swift
@@ -20,7 +20,7 @@ struct MovieEntity: Codable {
     let release_date: String
     let title: String
     let video: Bool
-    let vote_average: Int
+    let vote_average: Double
     let vote_count: Int
 }
